requirements.txt
Django>=3.2.7,<3.3.0
djangorestframework>=3.12.4,<3.13.0
python-decouple>=3.4,<3.5
flake8>=3.9.2,<3.10.0
Faker>=8.13.2,<8.14.0
psycopg2>=2.9.1,<2.10.0
Pillow>=8.3.2,<8.4.0
boto3>=1.18.63,<1.19.0
django_storages>=1.12.1,<1.13.0